+++ aiomoqt/client.py
@@ -9,7 +9,8 @@ from .protocol import (
     MOQTPeer, MOQTSessionQuic, MOQTSessionWTClient,
     DEFAULT_TX_MAX_INFLIGHT_BYTES,
 )
-from .types import moqt_alpn_for_version, moqt_version_from_draft, MOQTDraft
+from .types import (moqt_alpn_for_version, moqt_version_from_draft,
+                    MOQTDraft, require_d18_enabled)
 from .utils.logger import *
 
 logger = get_logger(__name__)
@@ -68,6 +69,7 @@ class MOQTClient(MOQTPeer):
             self.supported_drafts = tuple(supported_drafts)
         else:
             self.supported_drafts = (MOQTDraft.DRAFT_16, MOQTDraft.DRAFT_14)
+        require_d18_enabled(self.supported_drafts)
         self.draft_version = draft_version
         self.keylog_filename = keylog_filename
         self.configuration = configuration